Url query wrong mutation fix
Url selection can lead to URL mutating from `/endpoint` to `/endpoint?param=value` instead of `/endpoint/?param=value`

I'm not sure why is it parsing query parameters like that. It's not according to the !(RFC 3986)[https://datatracker.ietf.org/doc/html/rfc3986#section-4.2] and I'm aware of that. Currently, fixing it with this